Remove django requirement, turn into standard Python package
