+++ environment-dev.yml
@@ -2,7 +2,6 @@ name: gdalutils
 channels:
   - conda-forge
 dependencies:
-  - Django==3.2.7
   - GDAL=3.2.1
   - proj=7.1.1
   - pip:
