+++ gdal_utils/exceptions.py
