+++ requirements-dev.txt
@@ -1,9 +1,7 @@
 # Testing
 # ------------------------------------------------------------------------------
-django-stubs==1.9.0
 mypy==0.910
 pytest==6.2.5
-pytest-django==4.4.0
 requests==2.26.0
 requests-mock==1.9.3
 types-requests==2.25.11
@@ -14,8 +12,6 @@ autopep8==1.6.0
 beautifulsoup4==4.10.0 # Required for Fortify scanning.
 black==21.10b0
 coverage==6.1.1
-django-coverage-plugin==2.0.1
 flake8==4.0.1
 flake8-isort==4.1.1
-pylint-django==2.4.4
 pyupgrade==2.29.0
