+++ requirements.txt
@@ -1,3 +1,2 @@
-Django==3.2.7
 GDAL==3.2.1
 proj=7.1.1
