+++ setup.cfg
@@ -7,8 +7,6 @@ author = EventKit Contributors
 license = MIT
 classifiers =
     Environment :: Web Environment
-    Framework :: Django
-    Framework :: Django :: 3.2
     Intended Audience :: Developers
     License :: MIT License
     Operating System :: OS Independent
@@ -21,15 +19,12 @@ classifiers =
 include_package_data = true
 packages = find:
 install_requires =
-    Django >= 3.0.0, < 4.0.0
     GDAL==3.2.1
     MapProxy==1.13.2
 
 [coverage:run]
 include = auth/*
 omit = *migrations*, *tests*
-plugins =
-    django_coverage_plugin
 
 [flake8]
 max-line-length = 120
@@ -50,20 +45,9 @@ ignore_missing_imports = True
 warn_unused_ignores = True
 warn_redundant_casts = True
 warn_unused_configs = True
-plugins = mypy_django_plugin.main
-
-[mypy.plugins.django-stubs]
-django_settings_module = "config.settings"
-
-[mypy-*.migrations.*]
-# Django migrations should not produce any errors:
-ignore_errors = True
 
 [pylint.'FORMAT']
 good-names=a,i,j,k,pk
 
 [pylint.'MESSAGES CONTROL']
 disable = fixme, missing-class-docstring, missing-module-docstring, missing-function-docstring, too-many-arguments, unused-argument, too-few-public-methods, broad-except, too-many-locals, too-many-branches
-
-[tool:pytest]
-DJANGO_SETTINGS_MODULE = "config.settings"
